# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StudioFlow - Real Estate Media Delivery Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - in: query
          name: status
          schema:
            type: string
          required: false
        - in: query
          name: search
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get project media stats
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

  # Media
...
  /users/me:
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        businessName:
          type: ["string", "null"]
        plan:
          type: string
          enum: [free, pro, studio]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - plan
        - createdAt
...
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        businessName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
...
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        clientId:
          type: ["integer", "null"]
        name:
          type: string
        address:
          type: string
        propertyType:
          type: string
          enum: [residential, commercial, luxury, vacation, land]
        status:
          type: string
          enum: [draft, active, delivered, archived]
        coverImageUrl:
          type: ["string", "null"]
        mediaCount:
          type: integer
        shootDate:
          type: ["string", "null"]
        deliveryDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - name
        - address
        - propertyType
        - status
        - mediaCount
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        propertyType:
          type: string
          enum: [residential, commercial, luxury, vacation, land]
        clientId:
          type: ["integer", "null"]
        shootDate:
          type: ["string", "null"]
        deliveryDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - address
        - propertyType
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        propertyType:
          type: string
          enum: [residential, commercial, luxury, vacation, land]
        status:
          type: string
          enum: [draft, active, delivered, archived]
        clientId:
          type: ["integer", "null"]
        shootDate:
          type: ["string", "null"]
        deliveryDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
    ProjectStats:
      type: object
      properties:
        totalMedia:
          type: integer
        photos:
          type: integer
        videos:
          type: integer
        aiProcessed:
          type: integer
        approved:
          type: integer
        pending:
          type: integer
        totalSizeMb:
          type: number
        galleryViews:
          type: integer
        downloads:
          type: integer
      required:
        - totalMedia
        - photos
        - videos
        - aiProcessed
        - approved
        - pending
        - totalSizeMb
        - galleryViews
        - downloads
...