# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudioFlow - Real Estate Media Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectStats {
  totalMedia: number;
  photos: number;
  videos: number;
  aiProcessed: number;
  approved: number;
  pending: number;
  totalSizeMb: number;
  galleryViews: number;
  downloads: number;
}