# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudioFlow - Real Estate Media Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyPropertyType } from "./updateProjectBodyPropertyType";
import type { UpdateProjectBodyStatus } from "./updateProjectBodyStatus";

export interface UpdateProjectBody {
  name?: string;
  address?: string;
  propertyType?: UpdateProjectBodyPropertyType;
  status?: UpdateProjectBodyStatus;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  shootDate?: string | null;
  /** @nullable */
  deliveryDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
}